pkg/cvo/status: Filter out shallowly UpdateEffectNone errors from Failing condition
Various errors get propagated to users, such as the summarized task
graph error. For example, in the form of the message in the Failing
condition. However, update errors set with the update effect of
UpdateEffectNone can confuse users, as these primarily informing
messages get displayed together with valid update errors that heavily
impact the update. This can result in a message such as:

{
  "lastTransitionTime": "2023-06-20T13:40:12Z",
  "message": "Multiple errors are preventing progress:\n* Cluster
  operator authentication is updating versions\n* Could not update
  customresourcedefinition \"alertingrules.monitoring.openshift.io\"
  (512 of 993): the object is invalid, possibly due to local cluster
  configuration",
  "reason": "MultipleErrors",
  "status": "True",
  "type": "Failing"
}

The Failing condition is not true because of the UpdateEffectNone
error ("Cluster operator authentication is updating versions"), but
its message still gets displayed.

This commit makes sure that update errors that do not heavily affect
the update will be removed from the Failing condition message to an
extent.

This will improve the user experience. However, the removed errors
from the message will still be displayed in the logs and in other
places, such as the ReconciliationIssues condition.

The final error message that gets written to the Failing condition is a
tree of nodes (errors). To effectively remove all UpdateEffectNone
errors, it would be needed to remove all these nodes and update any
affected parent nodes, as the CVO creates aggregating errors. Currently,
it is not simple to robustly do this. It would be needed to recreate any
parent errors in the tree, and even this introduces new technical
challenges.

This commit aims to provide a simple way to remove any redundant errors
from the Failing condition message by checking the main error. If
applicable, it removes its immediate UpdateEffectNone nested errors and
recreates the main error. This approach has its limitations as well.
Any UpdateEffectNone errors deeper in the tree are ignored. However,
for the time being, this fix is sufficient and simple. Nested
MultipleErrors in MultipleErrors are not expected, however, in the
future, this commit may become less functional.

In the future, the overall logic regarding the error tree should be
addressed for functionalities such as this or for ReconciliationIssues
condition. For the time being, the commit approach represents a simple
way of improving the UX.